Specify EGL_PBUFFER_BIT when selecting EGL configuration.
Change the EGL config selection so that it sets the EGL_PBUFFER_BIT in the
EGL_SURFACE_TYPE attribute. The library makes use of pbuffers, so needs to
request a context that is guaranteed to support them.  Without this attribute,
the returned context may not support pbuffers and the application will fail.

Change-Id: I7ef408578e8d7ec58086c9baaf3a16371051fd79
Signed-off-by: jon.parr <jon.parr@imgtec.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>